[NUCCORE-1339] Use declaration type/metadata instead of instance class
Using the actual container class to obtain the ContainerHandler can lead
to a cast exception when we don't have the specific Java type definition
and a most compatible one that is returned based on the instance type is
not compatible with the field type. e.g Field is Multiset, container
class HashMultiset and the 
most compatible Java type returned is Collection, instead of Multiset.
So the JDKCollectionHandler is used which creates a new container of
type ArrayList by default, which turns out not to be assignable to
Multiset.